Compare Marape 3 to enwiki
I strongly suspect this isn't up-to-date even with this Cabinet (which
is no longer current anyway, as the 2nd Marape Government / 4th Marape
Cabinet is now in place).

There are a variety of reshuffles and changes listed at
https://fr.wikipedia.org/wiki/Gouvernement_Marape_I that don't appear to
be included here, so this should probably get a more precise end-date
earlier in the term.